add checkAnswer in utils
